feat: Remove unused or simply used function

+++ source/routes/medias/postMedia.controller.ts
@@ -1,17 +1,20 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 import { MultipartFile } from '@fastify/multipart';
 import { BadRequest, InternalServerHttpError, Unauthorized, UnsupportedMediaType } from '@library/httpError';
-import { execute, getHashFromStream, getMetadata, isValidType, readPartialFile, writeFileFromStream } from '@library/utility';
+import { getMetadata, isValidType } from '@library/utility';
 import { join } from 'path/posix';
-import { createReadStream } from 'fs';
+import { createReadStream, createWriteStream } from 'fs';
 import { prisma } from '@library/database';
 import { Media, MediaVideo, Prisma } from '@prisma/client';
-import { FileType, Metadata } from '@library/type';
+import { FileType, Metadata, RejectFunction, ResolveFunction } from '@library/type';
 import { deleteObjects, getObjectKeys, putObject } from '@library/bucket';
 import { ServiceOutputTypes } from '@aws-sdk/client-s3';
-import { readdir, rm } from 'fs/promises';
+import { FileHandle, open, readdir, rm } from 'fs/promises';
 import { mkdtemp } from 'fs/promises';
 import { tmpdir } from 'os';
+import { spawn } from 'child_process';
+import { read } from 'fs';
+import { Hash, createHash } from 'crypto';
 
 export default function (request: FastifyRequest, reply: FastifyReply): void {
 	/// @ts-expect-error :: imtoolazy
@@ -21,6 +24,7 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 		type: FileType;
 		isVideo: boolean;
 		hash: string;
+		mime: string;
 	} = {
 		isVideo: false
 	};
@@ -41,12 +45,20 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 				case 'mp4': {
 					if(request['user']['isVerified']) {
 						file['isVideo'] = true;
+						
+						break;
 					} else {
 						throw new Unauthorized('User must be verified');
 					}
 				}
-				case 'jpg':
+				case 'jpg': {
+					file['mime'] = 'image/jpeg';
+
+					break;
+				}
 				case 'png': {
+					file['mime'] = 'image/png';
+
 					break;
 				}
 
@@ -57,17 +69,58 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 
 			file['path'] = join(file['basePath'], 'input.' + file['type']);
 
-			return writeFileFromStream(file['path'], multipartFile['file']);
+			return new Promise<void>(function (resolve: ResolveFunction, reject: RejectFunction): void {
+				multipartFile['file'].pipe(createWriteStream(file['path']), {
+					end: true
+				})
+				.on('finish', function () {
+					resolve();
+					
+					return;
+				})
+				.once('error', reject);
+		
+				return;
+			});
 		} else {
 			throw new BadRequest('File must be exist');
 		}
 	})
 	.then(function (): Promise<Buffer> {
-		return readPartialFile(file['path'], 24);
+		return new Promise<Buffer>(function (resolve: ResolveFunction<Buffer>, reject: RejectFunction): void {
+			open(file['path'])
+			.then(function (fileHandle: FileHandle): void {
+				const buffer: Buffer = Buffer.alloc(24);
+	
+				read(fileHandle['fd'], buffer, 0, 24, 0, function (error: Error | null): void {
+					if(error === null) {
+						fileHandle.close()
+						.then(function (): void {
+							resolve(buffer);
+						})
+						.catch(reject);
+					} else {
+						reject(error);
+					}
+	
+					return;
+				});
+			})
+		});
 	})
 	.then(function (partialBuffer: Buffer): Promise<string> {
 		if(isValidType(partialBuffer, file['type'])) {
-			return getHashFromStream(createReadStream(file['path']));
+			return new Promise<string>(function (resolve: ResolveFunction<string>, reject: RejectFunction): void {
+				const hash: Hash = createHash('sha512').setEncoding('hex');
+		
+				createReadStream(file['path']).pipe(hash)
+				.once('error', reject);
+			
+				hash.once('finish', function (): void {
+					resolve(hash.read());
+				})
+				.once('error', reject);
+			});
 		} else {
 			throw new UnsupportedMediaType('File must be valid type');
 		}
@@ -104,9 +157,23 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 	} | null): Promise<void> | void {
 		if(media === null) {
 			if(file['isVideo']) {
-				return execute(`ffmpeg -v quiet -i input.mp4 -filter:v "scale='min(1280,iw)':min'(720,ih)':force_original_aspect_ratio=decrease,pad=1280:720:(ow-iw)/2:(oh-ih)/2" -map 0:v:0 -map 0:a:0 -c:v h264_qsv -c:a aac -q 23 -preset veryslow -f ssegment -segment_list "index.m3u8" "%d.ts"`, {
-					cwd: file['basePath'],
-					env: process['env']
+				return new Promise<void>(function (resolve: ResolveFunction, reject: RejectFunction): void {
+					spawn(`ffmpeg -v quiet -i input.mp4 -filter:v "scale='min(1280,iw)':min'(720,ih)':force_original_aspect_ratio=decrease,pad=1280:720:(ow-iw)/2:(oh-ih)/2" -map 0:v:0 -map 0:a:0 -c:v h264_qsv -c:a aac -q 23 -preset veryslow -f ssegment -segment_list "index.m3u8" "%d.ts"`, {
+						cwd: file['basePath'],
+						env: process['env'],
+						shell: true
+					})
+					.once('exit', function (code: number, error: Error | null): void {
+						if(error === null && code === 0) {
+							resolve();
+						} else {
+							reject(new Error('Process exited unexpectedly'));
+						}
+					
+						return;
+					});
+				
+					return;
 				});
 			} else {
 				return;
@@ -146,7 +213,7 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 				width: standardMetadata['video']['width'],
 				height: standardMetadata['video']['height'],
 				aspectRatio: standardMetadata['video']['aspectRatio'],
-				isVideo: false,
+				isVideo: true,
 				duration: 0,
 				frameRate: 0,
 				bitRate: 0,
@@ -155,8 +222,9 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 
 			for(let i: number = 0; i < paths['length']; i++) {
 				const filePath: string = join(file['basePath'], paths[i]);
+				const isVideo: boolean = paths[i].endsWith('ts');
 
-				if(paths[i].endsWith('ts')) {
+				if(isVideo) {
 					const metadata: Metadata = getMetadata(filePath, file['basePath']);
 
 					if(metadata['video']['bitRate'] < 0) {
@@ -180,7 +248,7 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 					(media['bitRate'] as NonNullable<typeof media['bitRate']>) += metadata['bitRate'];
 				}
 
-				putObjectPromises.push(putObject(join('videos', file['hash'], paths[i]), createReadStream(filePath)));
+				putObjectPromises.push(putObject(join('videos', file['hash'], paths[i]), createReadStream(filePath), isVideo ? 'video/MP2T' : 'application/x-mpegURL'));
 			}
 
 			(media['frameRate'] as NonNullable<typeof media['frameRate']>) /= meviaVideos['length'];
@@ -207,7 +275,7 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 				isVideo: false
 			};
 
-			putObjectPromises.push(putObject(join('images', file['hash'] + '.' + file['type']), createReadStream(file['path'])));
+			putObjectPromises.push(putObject(join('images', file['hash'] + '.' + file['type']), createReadStream(file['path']), file['mime']));
 		}
 		
 		return Promise.all(putObjectPromises);
@@ -230,7 +298,7 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 				width: true,
 				height: true,
 				aspectRatio: true,
-				isVideo: file['isVideo'],
+				isVideo: true,
 				duration: file['isVideo'],
 				frameRate: file['isVideo'],
 				bitRate: file['isVideo'],
@@ -247,20 +315,30 @@ export default function (request: FastifyRequest, reply: FastifyReply): void {
 			force: true,
 			recursive: true
 		})
-		.then(function (): Promise<string[]> {
-			return getObjectKeys('vidoes/' + file['hash'] + '/');
+		.then(function (): Promise<string[]> | string[] {
+			if(typeof(file['hash']) === 'string') {
+				return file['isVideo'] ? getObjectKeys('vidoes/' + file['hash'] + '/') : ['images/' + file['hash'] + '.' + file['type']];
+			} else {
+				return [];
+			}
 		})
-		.then(function (objectKeys: string[]): Promise<ServiceOutputTypes> {
-			console.log(objectKeys);
-			return deleteObjects(objectKeys);
+		.then(function (objectKeys: string[]): Promise<ServiceOutputTypes> | undefined {
+			if(objectKeys['length'] !== 0) {
+				return deleteObjects(objectKeys);
+			} else {
+				return;
+			}
 		})
-		.then(function (serviceOutput: ServiceOutputTypes): Promise<Prisma.BatchPayload> {
-			console.log(serviceOutput)
-			return prisma['media'].deleteMany({
-				where: {
-					hash: file['hash']
-				}
-			});
+		.then(function (): Promise<Prisma.BatchPayload> | undefined {
+			if(typeof(file['hash']) === 'string') {
+				return prisma['media'].deleteMany({
+					where: {
+						hash: file['hash']
+					}
+				});
+			} else {
+				return;
+			}
 		})
 		.then(function (): void {
 			reply.send(error);
